resources/stock/controller/app/requirements.txt
pylint==2.4.0
google-cloud==0.34.0
google-cloud-secret-manager==0.2.0
google-cloud-bigquery==1.24.0
google-resumable-media==0.5.0
Flask==1.1.2
kafka-python==2.0.1
gunicorn==20.0.4
cloudevents==0.2.4
requests==2.23.0
